refactor: move `isPointInPath` into core/mixin
